% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_cond_utils.R
\name{Adj_cond_utils}
\alias{Adj_cond_utils}
\title{Adjust the condition-specific utility for age and sex}
\usage{
Adj_cond_utils(
  Age_Sex_data,
  Condition_utilities,
  GenPopUtil = qalyr::GenPopUtil_2018_3L,
  Coefficients = qalyr::coefs_2018_3L
)
}
\arguments{
\item{Age_Sex_data}{mean age and proportion of males for in condition in data.
Calculated using function \code{MeanAge_Sex()}.}

\item{Condition_utilities}{condition-specific utilities.
Calculated using function \code{CalcUtils()}.}

\item{GenPopUtil}{table of general population utilities by age and sex.
Default stored in package as \code{qalyr::GenPopUtil_2018_3L}.}

\item{Coefficients}{vector of regression coefficients from the EQ-5D regression
model applied by Ara and Brazier. Default stored in package as \code{qalyr::coefs_2018_3L}}
}
\value{
Returns a summary table of condition-specific utilities.
}
\description{
Calculates an average utility for each alcohol/tobacco-related condition by age and sex.
Need to calculate the mean age and proportion of males for each condition in the HODAR data.
Need to calculate the matched age-sex specific general population utility using the \insertCite{Ara2010;textual}{qalyr} formula. Dividing these will give the multiplier by which to adjust the
general population utility score by, to get age-adjusted condition-specific utilities.
}
\examples{

\dontrun{

library(qalyr)

# Merge inpatient data and HODAR data and collapse to survey level data
data <- qalyr::merge_data(
  inpatient_data = qalyr::read_inpatient_data(),
  survey_data = qalyr::read_survey_data(),
  lkup = qalyr::lkup)

# Condition-specific utilities
cond_utils <- qalyr::CalcUtils(data, lkup = qalyr::lkup)

# Mean age and proportion of males
mean_age_sex <- qalyr::MeanAge_Sex(data, lkup = qalyr::lkup)

# age-sex-condition-specific utilities
adj_cond_utils <- qalyr::Adj_cond_utils(Age_Sex_data = mean_age_sex,
                                        Condition_utilities = cond_utils,
                                        GenPopUtil = qalyr::GenPopUtil)

}

}
\references{
\insertRef{Ara2010}{qalyr}
}
