% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcUtils.R
\name{CalcUtils}
\alias{CalcUtils}
\title{Calculate the average utility for each alcohol/tobacco-related condition}
\usage{
CalcUtils(data, lkup = qalyr::lkup)
}
\arguments{
\item{data}{}

\item{lkup}{table of conditions and respective ICD10 codes}
}
\value{
Returns a summary table of condition-specific utilities, including mean, standard deviation, and number of rows (to check for sample size).
}
\description{
Due to small sample sizes, we first use this function to calculate the average utility for each condition, without any demographic stratification.
Adj_cond_utils() is later used to stratify utilities by age and sex.
}
\examples{

\dontrun{

data <- merge_data(read_data(), lkup = qalyr::lkup)

condition_utilities <- CalcUtils(data = data, lkup = qalyr::lkup)

}

}
