% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefs}
\alias{coefs}
\title{Utility function coefficients (EQ-5D-3L)}
\format{
A named numeric vector
}
\usage{
coefs
}
\description{
[Ara and Brazier (2010)](https://www.sciencedirect.com/science/article/pii/S1098301510600903?via%3Dihub)
regression model coefficients with EQ-5D-3L utilities as the outcome and 
age, age squared, and sex as explanatory variables.
}
\keyword{datasets}
