% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefs_2018_5L}
\alias{coefs_2018_5L}
\title{Utility function coefficients 2018 (EQ-5D-5L)}
\format{
A named numeric vector
}
\usage{
coefs_2018_5L
}
\description{
Updated [Ara and Brazier (2010)](https://www.sciencedirect.com/science/article/pii/S1098301510600903?via%3Dihub)
regression model coefficients using Health Survey for England 2018 data with EQ-5D-5L utilities as the outcome and 
age, age squared, and sex as explanatory variables.
}
\keyword{datasets}
