% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenPopUtil}
\alias{GenPopUtil}
\title{General population utilities}
\format{
A data table (age, sex, GenPop_utility).
}
\source{
Published parameter values describing health state utility values by age and sex by \insertCite{Ara2010;textual}{qalyr}.
}
\usage{
GenPopUtil
}
\description{
General population utilities
}
\references{
\insertRef{Ara2010}{qalyr}
}
\keyword{datasets}
