% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenPopUtil_2018_5L}
\alias{GenPopUtil_2018_5L}
\title{General population utilities (EQ-5D-5L)}
\format{
A data table (age, sex, GenPop_utility)
}
\usage{
GenPopUtil_2018_5L
}
\description{
Parameters describing health state utility values by age and sex, originally 
estimated by [Ara and Brazier (2010)](https://www.sciencedirect.com/science/article/pii/S1098301510600903?via%3Dihub)
and updated with new analysis of Health Survey for England 2018 data.
}
\keyword{datasets}
