% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkup.R
\docType{data}
\name{lkup}
\alias{lkup}
\title{lkup data with alcohol and tobacco-related conditions and ICD-10 codes}
\format{A data table (icd_code, Description).}
\source{

}
\usage{
lkup
}
\description{
lkup data with alcohol and tobacco-related conditions and ICD-10 codes
}
\keyword{datasets}
