% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge the inpatient and survey data together}
\usage{
merge_data(inpatient_data, lkup)
}
\arguments{
\item{inpatient_data}{inpatient data containing data on the patient's admission and discharge dates, and diagnosis codes.}

\item{lkup}{lkup table in the format (icd_code, Description) (default stored in package as qalyr::lkup - this contains all
alcohol and tobacco related conditions).}
}
\value{
Returns a table with the survey data and inpatient data merged, with all admissions allocated to a survey,
ready for calculating condition-specific utilities.
}
\description{
As the inpatient data and survey data were provided separately, this function merges the two datasets, matching the survey information
to the inpatient data.
}
\details{
This function follows a set of rules to match the survey data to the inpatient data. The only field available to match the two datasets
is the Patient ID variable 'Patient_Number'. We have the date of admission and discharge in the inpatient data, and the date the survey
was sent and returned in the survey data.
}
\examples{

\dontrun{

inpatient_data <- read_data()

merge_data(inpatient_data, lkup = qalyr::lkup)

}

}
