% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qalyr-package.R
\docType{package}
\name{qalyr-package}
\alias{qalyr}
\alias{qalyr-package}
\title{qalyr: Tobacco- and Alcohol-Related Health State Utility Value Estimation}
\description{
Processes health state utility values from the HODaR data, and general population utility values from the Health Survey for England.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stapm.gitlab.io/r-packages/qalyr/}
}

}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Laura Webster
  \item Colin Angus \email{c.r.angus@sheffield.ac.uk}
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk}
}

}
\keyword{internal}
