% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reads in inpatient data}
\usage{
read_data(file = "D:/QALYs/Raw/inpatient_data.csv")
}
\value{
Returns a table of the inpatient data.
}
\description{
Reads in the inpatient data, retaining required columns: patient number, admission date, discharge date, episode number, diagnoses codes 1-13.
}
\examples{

\dontrun{

read_data()

}

}
