% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inpatient_data.R
\name{read_inpatient_data}
\alias{read_inpatient_data}
\title{Reads in inpatient HODaR data}
\usage{
read_inpatient_data(file = "D:/QALYs/Raw/inpatient_data.csv")
}
\arguments{
\item{file}{- the file path to the data}
}
\value{
Returns a table of the inpatient data.
}
\description{
Reads in the inpatient data, retaining required columns: patient number, admission date,
discharge date, episode number, diagnoses codes 1-13.
}
\examples{

\dontrun{

read_inpatient_data()

}

}
