% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey_data.R
\name{read_survey_data}
\alias{read_survey_data}
\title{Reads in survey HODaR data}
\usage{
read_survey_data(file = "D:/QALYs/Raw/routine_survey_data.csv")
}
\arguments{
\item{file}{- the file path to the data}
}
\value{
Returns a data.table of the survey data.
}
\description{
Reads in the survey data and calculate EQ5D.
}
\examples{

\dontrun{

read_survey_data()

}

}
