% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_cond_utils.R
\name{Adj_cond_utils}
\alias{Adj_cond_utils}
\title{Adjust the condition-specific utility for age and sex}
\usage{
Adj_cond_utils(
  Age_Sex_data = qalyr::Age_Sex_data,
  Condition_utilities = qalyr::util_data,
  GenPopUtil = qalyr::GenPopUtil
)
}
\arguments{
\item{Age_Sex_data}{mean age and proportion of males for in condition in data. Calculated using function *MeanAge_Sex()* - data stored in package
as qalyr::Age_Sex_data.}

\item{Condition_utilities}{condition-specific utilities. Calculated using function *CalcUtils()* - default stored in package as qalyr::util_data.}

\item{GenPopUtil}{table of general population utilities by age and sex. Default stored in package as qalyr::GenPopUtil}
}
\value{
Returns a summary table of condition-specific utilities.
}
\description{
Calculates an average utility for each alcohol/tobacco-related condition by age and sex.
Need to calculate the mean age and proportion of males for each condition in the HODAR data.
Need to calculate the matched age-sex specific general population utility using the Ara
and Brazier 2011 formula. Dividing these will give the multiplier by which to adjust the
general population utility score by, to get age-adjusted condition-specific utilities.
}
\examples{

\dontrun{


age_sex_condition_utilities <- Adj_cond_utils(Age_Sex_data = qalyr::Age_Sex_data, Condition_utilities = qalyr::util_data, GenPopUtil = qalyr::GenPopUtil)

}

}
