% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcUtils.R
\name{CalcUtils}
\alias{CalcUtils}
\title{Calculate the average utility for each alcohol/tobacco-related condition \lifecycle{maturing}}
\usage{
CalcUtils(data, lkup = qalyr::lkup)
}
\arguments{
\item{data}{Data.table - the merged HODaR inpatient and survey data}

\item{lkup}{Data table of conditions and respective ICD10 codes}
}
\value{
Returns a summary data table of condition-specific utilities,
including mean, standard deviation, and number of rows (to check for sample size).
}
\description{
Due to small sample sizes, we first use this function to calculate
the average utility for each condition, without any demographic stratification.
\code{Adj_cond_utils()} is later used to stratify utilities by age and sex.
}
\details{
For conditions where no utility data was available,
utilities were assumed to be similar to close conditions.
Thus, utilities for alcohol induced Cushing syndrome, alcoholic polyneuropathy and degeneration
were assumed to be similar to mental and behavioural disorders due to alcohol.
Utilities for psychosis were assumed to be similar to schizophrenia.
Utilities for alcoholic myopathy were assumed to be similar to utilities for alcohol cardiomyopathy.
The utility for excessive blood level of alcohol was assumed to be similar to evidence
of alcohol involvement determined by blood alcohol level.
Utilities for drowning were assumed to be similar to transport injuries.
Finally, utilities for other intentional injuries were assumed to be similar to other unintentional injuries.
}
\examples{

\dontrun{

data <- qalyr::merge_data(
  inpatient_data = qalyr::read_inpatient_data(),
  survey_data = qalyr::read_survey_data(),
  lkup = qalyr::lkup)

condition_utilities <- qalyr::CalcUtils(data = data, lkup = qalyr::lkup)

}

}
